ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
pillow>=11.0.0
pydantic>=2.10.0
certifi>=2024.0.0
